requirements.txt
autopep8==2.0.2
tensorflow==1.13.2
bchlib>=0.7
opencv-python>=3.4.3.18
numpy>=1.16.2
pillow>=5.4.1
stn>=1.0.1
autopep8
pycodestyle
pycryptodome
pyseltongue
pywatchman
scipy
six
sklearn
utilitybelt
